Refactor modals to separate files and make directories lowercase
