+++ public/pages/DetectorsList/Containers/ConfirmActionModals/__tests__/ConfirmDeleteDetectorsModal.test.tsx
@@ -16,11 +16,7 @@
 import React from 'react';
 import { render, fireEvent, wait } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import {
-  ConfirmStartDetectorsModal,
-  ConfirmStopDetectorsModal,
-  ConfirmDeleteDetectorsModal,
-} from '../ConfirmActionModals';
+import { ConfirmDeleteDetectorsModal } from '../ConfirmDeleteDetectorsModal';
 import { DetectorListItem, Monitor } from '../../../../../models/interfaces';
 import { DETECTOR_STATE } from '../../../../../utils/constants';
 
@@ -44,21 +40,6 @@ testMonitor['detector-id-0'] = [
   },
 ];
 
-const defaultStartProps = {
-  detectors: testDetectors,
-  hideModal: jest.fn(),
-  onStartDetectors: jest.fn(),
-  isListLoading: false,
-};
-
-const defaultStopProps = {
-  detectors: testDetectors,
-  monitors: {},
-  hideModal: jest.fn(),
-  onStopDetectors: jest.fn(),
-  isListLoading: false,
-};
-
 const defaultDeleteProps = {
   detectors: testDetectors,
   monitors: {},
@@ -67,88 +48,6 @@ const defaultDeleteProps = {
   onDeleteDetectors: jest.fn(),
   isListLoading: false,
 };
-
-describe('<ConfirmStartDetectorsModal /> spec', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-  describe('ConfirmStartDetectorsModal', () => {
-    test('renders modal with detectors', async () => {
-      const { getByText } = render(
-        <ConfirmStartDetectorsModal {...defaultStartProps} />
-      );
-      getByText('Are you sure you want to start the selected detectors?');
-      getByText('Start detectors');
-      getByText('detector-0');
-      getByText('detector-1');
-    });
-    test('should call onStartDetectors() when confirming', async () => {
-      const { getByTestId } = render(
-        <ConfirmStartDetectorsModal {...defaultStartProps} />
-      );
-      fireEvent.click(getByTestId('confirmButton'));
-      await wait();
-      expect(defaultStartProps.onStartDetectors).toHaveBeenCalled();
-    });
-    test('should call hideModal() when closing', async () => {
-      const { getByTestId } = render(
-        <ConfirmStartDetectorsModal {...defaultStartProps} />
-      );
-      fireEvent.click(getByTestId('cancelButton'));
-      await wait();
-      expect(defaultStartProps.hideModal).toHaveBeenCalled();
-    });
-  });
-});
-
-describe('<ConfirmStopDetectorsModal /> spec', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-  describe('ConfirmStopDetectorsModal', () => {
-    test('renders modal with detectors and no monitors', async () => {
-      const { getByText, getAllByText } = render(
-        <ConfirmStopDetectorsModal {...defaultStopProps} />
-      );
-      getByText('Are you sure you want to stop the selected detectors?');
-      getByText('Stop detectors');
-      getByText('detector-0');
-      getByText('detector-1');
-      expect(getAllByText('-')).toHaveLength(2);
-    });
-    test('renders modal with detectors and 1 monitor', async () => {
-      console.error = jest.fn();
-      const { getByText } = render(
-        <ConfirmStopDetectorsModal
-          {...defaultStopProps}
-          monitors={testMonitor}
-        />
-      );
-      getByText('Are you sure you want to stop the selected detectors?');
-      getByText('Stop detectors');
-      getByText('detector-0');
-      getByText('detector-1');
-      getByText('monitor-0');
-      getByText('-');
-    });
-    test('should call onStopDetectors() when confirming', async () => {
-      const { getByTestId } = render(
-        <ConfirmStopDetectorsModal {...defaultStopProps} />
-      );
-      fireEvent.click(getByTestId('confirmButton'));
-      await wait();
-      expect(defaultStopProps.onStopDetectors).toHaveBeenCalled();
-    });
-    test('should call hideModal() when closing', async () => {
-      const { getByTestId } = render(
-        <ConfirmStopDetectorsModal {...defaultStopProps} />
-      );
-      fireEvent.click(getByTestId('cancelButton'));
-      await wait();
-      expect(defaultStopProps.hideModal).toHaveBeenCalled();
-    });
-  });
-});
 
 describe('<ConfirmDeleteDetectorsModal /> spec', () => {
   beforeEach(() => {
